Added another auto template, and tried to fix some teleop!
