+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/IntoTheDeep24_25/auto/scrimmage/BlueLeft.java
@@ -1,20 +1,20 @@
-package org.firstinspires.ftc.teamcode.IntoTheDeep24_25.auto;
+package org.firstinspires.ftc.teamcode.IntoTheDeep24_25.auto.scrimmage;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.Servo;
-import org.firstinspires.ftc.teamcode.IntoTheDeep24_25.pidMaybe;
 
-@Autonomous(name = "BLUE RIGHT", group = "Autonomous")
-public class BlueRight extends LinearOpMode {
+@Autonomous(name = "BLUE LEFT", group = "Autonomous")
+public class BlueLeft extends LinearOpMode {
 
     private DcMotorEx frontRight, frontLeft, backRight, backLeft;
     private DcMotorEx armMotor = null;
     private Servo claw = null;
 
-    // Constants for movement
-    private final int TURN_DURATION = 500;
+    // Arm positions
+    private final int armUpPosition = 30;
+    private final int armDownPosition = 150;
 
     @Override
     public void runOpMode() throws InterruptedException {
@@ -36,58 +36,69 @@ public class BlueRight extends LinearOpMode {
 
         waitForStart();
 
-        // BLUE RIGHT Path
-        executePath();
-    }
+        // BLUE LEFT Path
+        // Turn left
+        mecanum(0, 0, -1); // RSX = -1 for left rotation
+        sleep(500); // Adjust timing for 90-degree turn
 
-    private void executePath() throws InterruptedException {
-        int MOVE_6_INCHES = 300;
-        int MOVE_28_INCHES = 1000;
-        int MOVE_95_INCHES = 3000;
-        int MOVE_100_INCHES = 3000;
-        int MOVE_105_INCHES = 3300;
-        int MOVE_85_INCHES = 2700;
-        int MOVE_36_INCHES = 1200;
-        int MOVE_5_INCHES = 200;
+        // Move forward 28 inches
+        mecanum(1, 0, 0); // Forward
+        sleep(1000); // Adjust timing for 28 inches
 
-        turnRight();
-        moveForward(MOVE_28_INCHES);
-        turnLeft();
-        moveForward(MOVE_6_INCHES);
+        // Turn right
+        mecanum(0, 0, 1); // RSX = 1 for right rotation
+        sleep(500); // Adjust timing for 90-degree turn
+
+        // Move forward 6 inches
+        mecanum(1, 0, 0); // Forward
+        sleep(300); // Adjust timing for 6 inches
+
+        // Grab Sample
         grabSample();
-        turnLeft();
-        moveForward(MOVE_95_INCHES);
-        turnLeft();
+
+        // Turn left
+        mecanum(0, 0, -1); // RSX = -1 for left rotation
+        sleep(500); // Adjust timing for 90-degree turn
+
+        // Place Basket
         placeSample();
-        turnLeft();
-        moveForward(MOVE_100_INCHES);
-        turnLeft();
+
+        // Turn right
+        mecanum(0, 0, 1); // RSX = 1 for right rotation
+        sleep(500); // Adjust timing for 90-degree turn
+
+        // Strafe left a little
+        mecanum(0, -1, 0); // LSX = -1 for left strafe
+        sleep(300); // Adjust timing for small strafe
+
+        // Move forward
+        mecanum(1, 0, 0); // Forward
+        sleep(1000); // Adjust timing for desired distance
+
+        // Grab Sample
         grabSample();
-        turnLeft();
-        moveForward(MOVE_105_INCHES);
+
+        // Turn
+        mecanum(0, 0, 1); // RSX = 1 for right rotation
+        sleep(500); // Adjust timing for 90-degree turn
+
+        // Place Sample
         placeSample();
-        turnLeft();
-        moveForward(MOVE_85_INCHES);
-        turnLeft();
-        moveForward(MOVE_36_INCHES);
-        turnLeft();
-        moveForward(MOVE_5_INCHES);
-        ascent();
-    }
 
-    private void turnLeft() throws InterruptedException {
-        mecanum(0, 0, -1);
-        sleep(TURN_DURATION);
-    }
+        // Turn left
+        mecanum(0, 0, -1); // RSX = -1 for left rotation
+        sleep(500); // Adjust timing for 90-degree turn
 
-    private void turnRight() throws InterruptedException {
-        mecanum(0, 0, 1);
-        sleep(TURN_DURATION);
-    }
+        // Move forward
+        mecanum(1, 0, 0); // Forward
+        sleep(1000); // Adjust timing for desired distance
 
-    private void moveForward(int duration) throws InterruptedException {
-        mecanum(1, 0, 0);
-        sleep(duration);
+        // Strafe left
+        mecanum(0, -1, 0); // LSX = -1 for left strafe
+        sleep(500); // Adjust timing for full strafe
+
+        // Level 1 Ascent
+        level1Ascent();
     }
 
     private void mecanum(double LSY, double LSX, double RSX) {
@@ -121,7 +132,6 @@ public class BlueRight extends LinearOpMode {
         telemetry.update();
 
         // Move arm down
-        int armDownPosition = 180;
         moveArm(armDownPosition);
 
         // Close claw
@@ -135,12 +145,10 @@ public class BlueRight extends LinearOpMode {
         telemetry.update();
 
         // Move arm up
-        // Arm positions
-        int armUpPosition = 30;
         moveArm(armUpPosition);
 
         // Open claw
-        claw.setPosition(0);
+        claw.setPosition(-1);
         sleep(1000); // Allow time for claw to release
     }
 
@@ -149,9 +157,9 @@ public class BlueRight extends LinearOpMode {
         armMotor.setTargetPosition(targetPosition);
         armMotor.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
 
-       // pidMaybe pid = new pidMaybe(0.004, 0, armMotor);
+        //pidMaybe pid = new pidMaybe(0.004, 0, armMotor);
         while (opModeIsActive() && Math.abs(armMotor.getCurrentPosition() - targetPosition) > 5) {
-           // double power = pid.calculatePower(targetPosition, armMotor.getCurrentPosition());
+            //double power = pid.calculatePower(targetPosition, armMotor.getCurrentPosition());
             double power = 0;
             armMotor.setPower(power);
 
@@ -164,12 +172,10 @@ public class BlueRight extends LinearOpMode {
         armMotor.setPower(0); // Stop motor when done
     }
 
-    // Raise the arm to its ascent position
-    private void ascent() {
-        int ascentPosition = 200;
-        telemetry.addData("Action", "Ascent");
-        telemetry.update();
-        moveArm(ascentPosition);
-        sleep(500);
+    // Level 1 Ascent
+    private void level1Ascent() {
+        int level1AscentPosition = 200; // Adjust this value for Level 1 height
+        moveArm(level1AscentPosition);
+        sleep(500); // Ensure time for arm to stabilize
     }
 }
